refactor: change behavior to a more flexible way of create grids
To improve templating and allow custom elements (such as header and footer), the library has been
completely reviewed. Improvements have been made to the way it manage pages to be more simple and
less open to bugs (less is more !).

BREAKING CHANGE: the library does not work AT ALL as older versions. Consider to change your app or
it will generate errors.