core: new templates and generation process
The structure is now more granulary to improve flexibility and precisions. Tasks are more
shared between classes to prevent a single-class centered process. That allow more efficient
bug tracking, fixing and improvements. The template is now more disconected to the base
process wich reduce the creation of errors.

BREAKING CHANGE: the way to create templates and generate pdf has changed